+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog
 0.10.dev8
 ---------
 
+- added data-context-url for sane cartData and validate_cart_item calls on Plone 5
+  [agitator]
+
 - Plone 5 update
   [agitator]